+++ hutool-http/src/main/java/cn/hutool/http/HttpConnection.java
@@ -349,11 +349,13 @@ public class HttpConnection {
 	 * 采用流方式上传数据，无需本地缓存数据。<br>
 	 * HttpUrlConnection默认是将所有数据读到本地缓存，然后再发送给服务器，这样上传大文件时就会导致内存溢出。
 	 * 
-	 * @param blockSize 块大小（bytes数）
+	 * @param blockSize 块大小（bytes数），0或小于0表示不设置Chuncked模式
 	 * @return this
 	 */
 	public HttpConnection setChunkedStreamingMode(int blockSize) {
-		conn.setChunkedStreamingMode(blockSize);
+		if(blockSize > 0) {
+			conn.setChunkedStreamingMode(blockSize);
+		}
 		return this;
 	}